chromium-wayland: fix compilation problem on non-X11 builds
Certain build configurations enable a code path that includes
keyboard_code_conversion_xkb.cc, but it eventually fails in non-X11
environments because of a missing patch that is present upstream.

This commit adds the missing patch from
https://codereview.chromium.org/2239833003

There are two versions of the patch in the code review URL above.
The first version fixes the problem both under the gyp and GN build
systems. That patch was rejected because gyp is considered deprecated
upstream. The second version fixes the problem only on GN and it was
merged upstream. We are picking the first version because our recipes
still use gyp to build.

Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>